Update to hackaton-2022-radek-barton-timeline-view branch.

+++ src/PerfView/StackViewer/TimelineVisuals.cs
@@ -5,9 +5,9 @@ namespace PerfView
 {
     internal class TimelineVisuals
     {
-        public int StartingFrame { get; set; }
+        public float StartingFrame { get; set; }
 
-        public int EndingFrame { get; set; }
+        public float EndingFrame { get; set; }
 
         public Dictionary<int, List<WorkVisual>> VisualsPerThreadId { get; set; } = new Dictionary<int, List<WorkVisual>>();
     }
@@ -17,8 +17,8 @@ namespace PerfView
     /// </summary>
     internal class WorkVisual
     {
-        public int StartingFrame { get; set; }
-        public int EndingFrame { get; set; }
+        public float StartingFrame { get; set; }
+        public float EndingFrame { get; set; }
         public string DisplayName { get; set; }
         public Color DisplayColor { get; set; }
